+++ InformedProteomics.Backend/Data/Spectrometry/Spectrum.cs
@@ -435,7 +435,7 @@ namespace InformedProteomics.Backend.Data.Spectrometry
             for (var i = 0; i < observedPeaks.Length; i++)
             {
                 var observedPeak = observedPeaks[i];
-                observedIntensities[i] = observedPeak != null ? (float)observedPeak.Intensity : 0.0;
+                observedIntensities[i] = observedPeak != null ? observedPeak.Intensity : 0.0;
             }
             return FitScoreCalculator.GetPearsonCorrelation(isotopomerEnvelope, observedIntensities);
         }
@@ -488,7 +488,7 @@ namespace InformedProteomics.Backend.Data.Spectrometry
             {
                 theoIntensities[i] = isotopomerEnvelope[i];
                 var observedPeak = observedPeaks[i];
-                observedIntensities[i] = observedPeak != null ? (float)observedPeak.Intensity : 0.0;
+                observedIntensities[i] = observedPeak != null ? observedPeak.Intensity : 0.0;
             }
             return FitScoreCalculator.GetCosine(isotopomerEnvelope, observedIntensities);
         }
